Crate steel_macros

Steel Macros

Macros for the Steel Minecraft server.

Modules

behavior 🔒

packet 🔒

read_from 🔒

strategy 🔒

write_to 🔒

Attribute Macros

block_behavior

Marks a struct as a block behavior for auto-registration.

entity_behavior

Marks a struct as an entity implementation for auto-registration.

item_behavior

Marks a struct as an item behavior for auto-registration.

Derive Macros

ClientPacket

Derives the ClientPacket trait for a struct.

ReadFrom

Derives the ReadFrom trait for a struct.

ServerPacket

Derives the ServerPacket trait for a struct.

WriteTo

Derives the WriteTo trait for a struct.